- Buttons links.

+++ src/app/tag-hierarchy/tag-hierarchy-list/tag-hierarchy-list.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { TagHierarchy } from '../tag-hierarchy';
+import { ModalService } from './../../shared/confirm-modal/modal.service';
+import { ConfirmModalComponent } from './../../shared/confirm-modal/confirm-modal.component';
 
 @Component({
   selector: 'app-tag-hierarchy-list',
@@ -11,12 +13,22 @@ export class TagHierarchyListComponent implements OnInit {
 
   @Output() public deleteItem: EventEmitter<number> = new EventEmitter<number>();
 
-  constructor() { }
+  constructor(private confirmService: ModalService) { }
 
   ngOnInit() {
   }
 
-  delete(index: number): void {
-    this.deleteItem.emit(index);
+  public delete(index: number, $event): void {
+    this.confirmService.init(ConfirmModalComponent, {
+      title: 'Delete tag hierarchy',
+      message: 'Delete tag hierarchy?'
+    }).subscribe(
+      deleted => {
+        if (deleted) {
+          this.deleteItem.emit(index);
+          $event.stopPropagation();
+        }
+      }
+    );
   }
 }
